Switch from delay to wait function.
Wait will pause for a given number of milliseconds. This function sleeps the process to share the processor with other programs. A program that waits for even a few milliseconds will consume very little processor time. It is slightly less accurate than the pygame.time.delay.  However delay will continue using the processor to 100%.